[IMP] hr: Allow plans to directly request signatures
