+++ addons/hr/models/__init__.py
@@ -9,6 +9,7 @@ from . import hr_department
 from . import hr_departure_reason
 from . import hr_job
 from . import hr_plan
+from . import hr_plan_activity_type
 from . import hr_work_location
 from . import mail_alias
 from . import mail_channel
