+++ addons/hr_contract/data/hr_contract_demo.xml
@@ -85,6 +85,10 @@
         <field name="kanban_state">normal</field>
     </record>
 
+    <record id="hr.employee_qdp" model="hr.employee">
+        <field name="contract_id" ref="hr_contract.hr_contract_qdp"/>
+    </record>
+
     <record id="hr_contract_han" model="hr.contract">
         <field name="name">Walter Horton Contract</field>
         <field name="date_start" eval="time.strftime('%Y')+'-3-1'"/>
